planning error fixed

+++ Nav Apparel/CardPages/Planning/src/Pag50100.NetronicVSDemoApp.al
@@ -1298,6 +1298,13 @@ page 50324 "NETRONICVSDevToolDemoAppPage"
                                                 TargetPerHour := round(TargetPerDay / HoursPerDay, 1);
                                                 TempQty := 0;
 
+                                                //Delete old lines
+                                                ProdPlansDetails.Reset();
+                                                ProdPlansDetails.SetRange("Line No.", LineNo);
+                                                ProdPlansDetails.SetFilter(ProdUpd, '=%1', 0);
+                                                if ProdPlansDetails.FindSet() then
+                                                    ProdPlansDetails.DeleteAll();
+
                                                 repeat
 
                                                     //Get working hours for the day
@@ -1327,6 +1334,50 @@ page 50324 "NETRONICVSDevToolDemoAppPage"
 
                                                         if Count < 14 then
                                                             Error('Calender is not setup for the Line : %1', ResourceName);
+                                                    end;
+
+
+                                                    ResourceRec.Reset();
+                                                    ResourceRec.SetRange("No.", ResourceNo);
+                                                    ResourceRec.FindSet();
+
+                                                    if HoursPerDay = 0 then begin
+
+                                                        //Validate the day (Holiday or Weekend)
+                                                        SHCalHolidayRec.Reset();
+                                                        SHCalHolidayRec.SETRANGE("Shop Calendar Code", ResourceRec."Shop Calendar Code");
+                                                        SHCalHolidayRec.SETRANGE(Date, TempDate);
+
+                                                        if not SHCalHolidayRec.FindSet() then begin  //If not holiday
+                                                            DayForWeek.Get(DayForWeek."Period Type"::Date, TempDate);
+
+                                                            case DayForWeek."Period No." of
+                                                                1:
+                                                                    Day := 0;
+                                                                2:
+                                                                    Day := 1;
+                                                                3:
+                                                                    Day := 2;
+                                                                4:
+                                                                    Day := 3;
+                                                                5:
+                                                                    Day := 4;
+                                                                6:
+                                                                    Day := 5;
+                                                                7:
+                                                                    Day := 6;
+                                                            end;
+
+                                                            SHCalWorkRec.Reset();
+                                                            SHCalWorkRec.SETRANGE("Shop Calendar Code", ResourceRec."Shop Calendar Code");
+                                                            SHCalWorkRec.SetFilter(Day, '=%1', Day);
+                                                            if SHCalWorkRec.FindSet() then   //If not weekend
+                                                                Error('Calender for date : %1  Work center : %2 has not calculated', TempDate, ResourceRec.Name)
+                                                            else
+                                                                Holiday := 'Yes';
+                                                        end
+                                                        else
+                                                            Holiday := 'Yes';
                                                     end;
 
                                                     //No learning curve for holidays
@@ -1448,7 +1499,16 @@ page 50324 "NETRONICVSDevToolDemoAppPage"
 
                                                     ProdPlansDetails.Qty := xQty;
                                                     ProdPlansDetails.Target := TargetPerDay;
-                                                    ProdPlansDetails.HoursPerDay := HoursPerDay;
+                                                    // ProdPlansDetails.HoursPerDay := HoursPerDay;
+
+                                                    if Holiday = 'NO' then
+                                                        if TempHours > 0 then
+                                                            ProdPlansDetails.HoursPerDay := TempHours
+                                                        else
+                                                            ProdPlansDetails.HoursPerDay := HoursPerDay
+                                                    else
+                                                        ProdPlansDetails.HoursPerDay := 0;
+
                                                     ProdPlansDetails.ProdUpd := 0;
                                                     ProdPlansDetails.ProdUpdQty := 0;
                                                     ProdPlansDetails."Created User" := UserId;
@@ -1739,7 +1799,8 @@ page 50324 "NETRONICVSDevToolDemoAppPage"
                         ProdPlansDetails.Reset();
                         ProdPlansDetails.SetRange("Line No.", LineNo);
                         ProdPlansDetails.SetFilter(ProdUpd, '=%1', 0);
-                        ProdPlansDetails.DeleteAll();
+                        if ProdPlansDetails.FindSet() then
+                            ProdPlansDetails.DeleteAll();
 
                         //Check learning curve                        
                         LCurveFinishDate := dtStart;
@@ -2334,11 +2395,24 @@ page 50324 "NETRONICVSDevToolDemoAppPage"
                                             TargetPerHour := round(TargetPerDay / HoursPerDay, 1);
                                             TempQty := 0;
 
+                                            //Delete old lines
+                                            ProdPlansDetails.Reset();
+                                            ProdPlansDetails.SetRange("Line No.", LineNo);
+                                            ProdPlansDetails.SetFilter(ProdUpd, '=%1', 0);
+                                            if ProdPlansDetails.FindSet() then
+                                                ProdPlansDetails.DeleteAll();
+
                                             repeat
+
+                                                ResourceRec.Reset();
+                                                ResourceRec.SetRange("No.", ResourceNo);
+                                                ResourceRec.FindSet();
 
                                                 //Get working hours for the day
                                                 HoursPerDay := 0;
                                                 Rate := 0;
+                                                Holiday := 'No';
+
                                                 WorkCenCapacityEntryRec.Reset();
                                                 WorkCenCapacityEntryRec.SETRANGE("No.", ResourceNo);
                                                 WorkCenCapacityEntryRec.SETRANGE(Date, TempDate);
@@ -2363,6 +2437,45 @@ page 50324 "NETRONICVSDevToolDemoAppPage"
 
                                                     if Count < 14 then
                                                         Error('Calender is not setup for the Line : %1', ResourceName);
+                                                end;
+
+                                                if HoursPerDay = 0 then begin
+
+                                                    //Validate the day (Holiday or Weekend)
+                                                    SHCalHolidayRec.Reset();
+                                                    SHCalHolidayRec.SETRANGE("Shop Calendar Code", ResourceRec."Shop Calendar Code");
+                                                    SHCalHolidayRec.SETRANGE(Date, TempDate);
+
+                                                    if not SHCalHolidayRec.FindSet() then begin  //If not holiday
+                                                        DayForWeek.Get(DayForWeek."Period Type"::Date, TempDate);
+
+                                                        case DayForWeek."Period No." of
+                                                            1:
+                                                                Day := 0;
+                                                            2:
+                                                                Day := 1;
+                                                            3:
+                                                                Day := 2;
+                                                            4:
+                                                                Day := 3;
+                                                            5:
+                                                                Day := 4;
+                                                            6:
+                                                                Day := 5;
+                                                            7:
+                                                                Day := 6;
+                                                        end;
+
+                                                        SHCalWorkRec.Reset();
+                                                        SHCalWorkRec.SETRANGE("Shop Calendar Code", ResourceRec."Shop Calendar Code");
+                                                        SHCalWorkRec.SetFilter(Day, '=%1', Day);
+                                                        if SHCalWorkRec.FindSet() then   //If not weekend
+                                                            Error('Calender for date : %1  Work center : %2 has not calculated', TempDate, ResourceRec.Name)
+                                                        else
+                                                            Holiday := 'Yes';
+                                                    end
+                                                    else
+                                                        Holiday := 'Yes';
                                                 end;
 
                                                 //No learning curve for holidays
@@ -2484,7 +2597,16 @@ page 50324 "NETRONICVSDevToolDemoAppPage"
 
                                                 ProdPlansDetails.Qty := xQty;
                                                 ProdPlansDetails.Target := TargetPerDay;
-                                                ProdPlansDetails.HoursPerDay := HoursPerDay;
+                                                // ProdPlansDetails.HoursPerDay := HoursPerDay;
+
+                                                if Holiday = 'NO' then
+                                                    if TempHours > 0 then
+                                                        ProdPlansDetails.HoursPerDay := TempHours
+                                                    else
+                                                        ProdPlansDetails.HoursPerDay := HoursPerDay
+                                                else
+                                                    ProdPlansDetails.HoursPerDay := 0;
+
                                                 ProdPlansDetails.ProdUpd := 0;
                                                 ProdPlansDetails.ProdUpdQty := 0;
                                                 ProdPlansDetails."Created User" := UserId;
